Explicitly end crashtracking job that hangs (#6955)
## Summary of changes

Add a timeout to the crash-tracking smoke test job

## Reason for change

This job regularly hangs, and we don't know why, because when it does,
the job times out, and we lose all the logs.

## Implementation details

Add a timeout to the _task_ so that the job can upload the logs and give
us a shot at fixing it.

## Test coverage

This is the test

## Other details

I don't know why we didn't do this before 🙈