Use a curled version of the dotnet-install script (#6949)
## Summary of changes

Instead of using the vendored dotnet-install script, `curl` a specific
version

## Reason for change

Using the local version was breaking the docker build cache in CI,
because when we do a `git clone` the timestamps on the file have
changed. This causes a large chunk of the pre-building of docker images
to be a waste of time, as nothing is cached.

We considered the following options:
:one: Stop using the copied dotnet-install.sh script
:two: Stop using Docker
:three: Pre-build Docker build images and run them directly, without
building in CI
:four: "Fix" the copy issue by resetting timestamps and ownership (using
`rsync`/`find`/`chmod`)
:five: "Fix" the copy issue by using a `tar` stream input to the `docker
build` step

And opted for 1️⃣, this PR. 

## Implementation details

- Revert the PR (mostly) that introduced the vendored scripts
- Replace the generic "latest" version of the script with a specific
github-sha version to ensure consistency

## Test coverage

This is the test. If the build works, we can update the VMs and make
sure the caching issue is resolved

## Other details

There's a minor downside in cases where we _don't_ have caching (mostly
the smoke tests). Is it worth keeping it vendored for those, and
avoiding hitting GitHub? 🤔 I'm torn between consistency and potential
reliability...