Allow disabling remote config management explicitly (#6942)
## Summary of changes

Add `DD_REMOTE_CONFIGURATION_ENABLED` setting

## Reason for change

There are some situations in which remote config is not available -
basically whenever the full agent isn't enabled. Currently we calculate
this dynamically, but this allows explicitly disabling it.

The main need is that the smoke tests _don't_ support RCM, and we're
seeing some flake due to the tracer adding a warning in the logs, and
our log scanner failing the build on this basis. Instead, we should
disable RCM as we know it's not available.

Note that this config value is already available in other languages:
ruby, python, js... I stopped checking after that 😅

## Implementation details

- Add a new config, use it to override the value of
`IsRemoteConfigurationAvailable`
- Disable remote config in all the smoke tests

## Test coverage

Added unit tests for the value of `IsRemoteConfigurationAvailable`

## Other details

PR for adding the config to intake:
- https://github.com/DataDog/dd-go/pull/182611