Disable Kafka header injection when brokers return unknown error (#6945)
## Summary of changes

This will disable our header injection for the Kafka instrumentation
when a message is produced and we detect that the message format used in
the broker may not support headers in messages. This happens when the
message format used is prior to 0.11.

## Reason for change

For Kafka message format versions before 0.11 they do not support
headers in messages, yet we always try to inject headers in the
messages.
When we hit this case the broker rejects the message and an unknown
broker error is returned, but since header injection works fine it never
gets disabled as the API supports the headers, but not the broker.

This checks any exception after producing a message for `Unknown broker
error` and if it is detected we then _assume_ that this is caused by our
header injection and a topic that was created with a message format
before 0.11 and then disables future header injection globally.

## Implementation details

## Test coverage

Unsure how to create a test for this to run in CI or if it is really
worthwhile to do->

```text
[DBG] Span closed: [s_id: 59bb8f7c569c5d48, p_id: null, t_id: 681cb83800000000f364d9d830bb265b] for (Service: Samples.Kafka-kafka, Resource: Produce Topic sample-topic, Operation: kafka.produce, Tags: [error.msg (tag):Unknown broker error,error.type (tag):System.Exception,error.stack (tag):System.Exception: Unknown broker error
```

With this change we _still_ get those errors, just after the "first" one
is detected we then disable header injection:

```text
[ERR] Kafka Broker responded with UNKNOWN_SERVER_ERROR (-1). Please look at broker logs for more information. Tracer message header injection for Kafka is disabled. Confluent.Kafka.ProduceException`2[System.String,System.String]: Unknown broker error
```

The sample application though runs many produce messages at the same
time, this first batch will still fail, but after we disable it the next
set of produce message calls succeed (without header injection).

I had to strip many of the `environment` stuff as with them they were
throwing tons of errors in the containers.

```docker
  kafka-zookeeper:
    image: confluentinc/cp-zookeeper:3.3.0
    hostname: kafka-zookeeper
    container_name: kafka-zookeeper
    ports:
      - "2181:2181"
    restart: unless-stopped
    environment:
      ZOOKEEPER_CLIENT_PORT: 2181
      ZOOKEEPER_TICK_TIME: 2000

  kafka-broker:
    image: bitnami/kafka:2.4.0
    hostname: kafka-broker
    container_name: kafka-broker
    depends_on:
      - kafka-zookeeper
    ports:
      - "9092:9092"
      - "9101:9101"
    restart: unless-stopped
    environment:
      KAFKA_BROKER_ID: 1
      KAFKA_ZOOKEEPER_CONNECT: kafka-zookeeper:2181
      KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://localhost:9092
      KAFKA_LISTENERS: PLAINTEXT://0.0.0.0:9092
      KAFKA_CFG_LOG_MESSAGE_FORMAT_VERSION: 0.10.2
      ALLOW_PLAINTEXT_LISTENER: yes
```

## Other details

We do not have the ability to query the Kafka broker version.
The messages with the injected headers will fail, but we decided that we
won't be retrying these messages without the headers.
Fixes https://datadoghq.atlassian.net/browse/DSMON-668

If
[devlieryReports](https://github.com/confluentinc/confluent-kafka-dotnet/blob/65362199f13bdad8a0831541f53d92e1e95a8a37/src/Confluent.Kafka/Producer.cs#L869)
are not enabled we won't be able to handle the case unfortunately - by
default delivery reports are enabled and we think this is a rare issue.

I followed the error message in
https://github.com/DataDog/dd-trace-py/pull/13313/
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->