Remove `isBenchmarksOnlyBuild` from the pipeline (#6897)
## Summary of changes

Remove the logic / code for benchmark only builds of the tracer as we
won't be needing it soon (and it is currently failing).

## Reason for change

We are migrating to the benchmarking platform so this won't be needed
soon.
Also it fails at the moment as the old branches aren't being properly
cleaned up / archived for some reason.

## Implementation details

- Removes `auto_update_benchmark_branches.yml` which archived the old
benchmark branch and created a new one
- Removes `benchmark` branch(es) from the daily build schedule
- Removes the `isBenchmarksOnlyBuild` logic from the pipeline.
- Removes the `isMainRepository` logic from the pipeline as this was
unused and leftover from before debugger was merged into this repo.

## Test coverage

- This is the test I'd guess


## Other details

Probably should go through the `benchmarks` branches left over and
archive them?
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->