Fix arm64 SQL server not found flaky tests (#6948)
## Summary of changes

In some alpine, arm64 tests, some SqlServer tests are failing because
the server is not found.

This PR adds the skip (13) return code to the samples, so we can skip
the test and log it.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->