[Dynamic Instrumentation] DEBUG-3796 Fix rate limiting for probes that don't have limit (#6921)
## Summary of changes
Fix issue that cause span decoration probes to not proceed due a
sampler.

## Implementation details
Span decoration probes (also metric probe) should not use the default
rate limit. Currently, we don't add any limit when creating the probe,
but during the processing of the probe, we are trying the get the probe
metadata for that specific probe instance and when we can't find
sampling, we are creatin gone with default limit. This PR fix that by
providing a not limit sampler for those probes.