Update Windows to hosted managed DevOps pool (#6960)
## Summary of changes

Move the Windows VMs from a VMSS pool to use a Hosted Managed DevOps
Pool

## Reason for change

We want to unify the infrastructure hence the move from VMSS to hosted
pools

## Implementation details

- Set up a new hosted devops pool (I will document this process in
confluence once I'm/we're happy)
- Reused the same image, so should have identical behaviour otherwise
- Update ultimate-pipeline to use the new pool

## Test coverage

This is the test

## Other details

Supersedes
- https://github.com/DataDog/dd-trace-dotnet/pull/6819

As it's a new pool, we have to _manually_ specify pre-provisioning. I've
currently set this to 9-5 UTC with 10VMs but we can tweak that as we see
fit. It's currently just to ensure we're getting some pre-provisioning
benefit. We can tweak that as we see fit (and can use predictive mode in
~3 weeks time, once Azure has time to analyze our usage)