Add debug output to crashtracking (#6973)
## Summary of changes

Add `DD_CRASHTRACKING_INTERNAL_LOG_TO_CONSOLE` to display debug messages
to the console in crashtracking.

## Reason for change

Crashtracking is painful to debug (citation needed), hopefully this
should help understand why it sometimes fail in the CI.

## Implementation details

Just writing stuff to the console when the environment variable is set.
Nothing fancy.

## Other details

Keep in mind that this is only useful for Linux. On Windows,
crashtracking runs in WerFault.exe which is not an interactive process
(so no console).

As a bonus, I cleaned up some minor stuff (moved the environment
variable names to constants).