Revert "Revert Load the tracer/profiler after guardrails checks" (#6959)
Reverts https://github.com/DataDog/dd-trace-dotnet/pull/6200

It was crashing in mixed-runtimes scenario because we added a safeguard
against double-initialization in the dispatcher, which incidentally
caused the tracer to be initialized twice (which, it turns out, it
doesn't like). Fixed by making sure each runtime gets its own instance
of the dispatcher.