Update Linux x64 to hosted managed DevOps pool (#6966)
## Summary of changes

- Switch to using Managed DevOps pools
- Fix caching for latest updates

## Reason for change

We want to unify the infrastructure hence the move from VMSS to hosted
pools

## Implementation details

- Set up a new hosted devops pool (documentation in-progress)
- Rebuilt the images to fix caching
- Update `ultimate-pipeline` to use the new pool

## Test coverage

This is the test

## Other details

As it's a new pool, we have to manually specify pre-provisioning. I've
currently set this to 9-5 UTC with 10VMs but we can tweak that as we see
fit. It's currently just to ensure we're getting some pre-provisioning
benefit. We can tweak that as we see fit (and can use predictive mode in
~3 weeks time, once Azure has time to analyze our usage)