Try to work around crash tracking errors (#6977)
## Summary of changes

Try some things to avoid hanging in crash tracking tests

## Reason for change

We see hanging in the crash tracking sometimes, and based on newly added
debug logs, it seems to happen _after_ `dd-dotnet` has exited. That
means our options are basically:

- either dd-dotnet did not in fact exit properly (but that would be very
surprising since we're printing the "exit" message at the very end)
- or it did not correctly resume the threads of the crashing app
- or it's not crashtracking that is causing the timeout

I suspect the answer is 2, but ChatGPT suggested these changes as
possible workarounds for niche issues for the final option, so trying
them.

Also, we're currently running the crash test when the job has already
failed, when we shouldn't be

## Implementation details

- Clean up the docker image
- pipe stderr as well as stdout
- Fix `condition`

## Test coverage

If this test passes, that's good enough, and we'll just keep an eye out
to see if it keeps happening anyway