Avoid gitlab jobs failing due to timeouts or interruptions (#6981)
## Summary of changes

Some gitlab benchmark jobs are failing in master either due to timeouts
or cancellation (because a new pipeline is started). We don't want the
CI to interrupt these jobs and fail, that's why the timeout has been
increased and the jobs are set to not interruptible (default value).

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->