[Profiler] Provide exact minor.build version information for .NET Framework applications (#6974)
## Summary of changes
Compute the minor and build parts of the .NET Framework version instead
of always returning 4.0
 
## Reason for change
It is important to make the difference between minor.build versions for
Call To Action part of the UI when a recommendation is needed for a
specific profile type

## Implementation details
Look at mscorlib.dll file version information and infer the build part
from the documented range of minimum versions from 4.6.1 to 4.8.1

## Test coverage
N/A

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->