Product mask support for DSM (#6961)
## Summary of changes
DSM backend will route payloads differently based on how the client
which produced them was configured. This change mirrors the one [done
for Java ](https://github.com/DataDog/dd-trace-java/pull/8051).
Ticket [DSMON-778](https://datadoghq.atlassian.net/browse/DSMON-778).

## Reason for change
This flag will allow us to better understand in which environments and
with which configurations DSM is being run. The value of the tag is a
bitmask which can be extended as needed in the future.

## Implementation details
The DSM payload was extended with `ProductMask` field. 

## Test coverage
Existing tests were updated.


[DSMON-778]:
https://datadoghq.atlassian.net/browse/DSMON-778?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ