Fix a couple of bugs on the instrumentation generator (#6985)
## Summary of changes

Fixing some bugs I encountered when working on my latest integrations:
 - when instrumenting interfaces, the instrumentation kind was not set
- integration name was generated using the method name, which is never
what we want. I replaced it with a hypothetical integration ID bearing
the name of the DLL, which (1) has chances of being correct and (2) will
be red in the editor if it's not the right one (better than having
everything compile and not work without knowing why)
- do not suggest type constraints when the instrumented type is generic
(see
[limitations](https://github.com/DataDog/dd-trace-dotnet/blob/master/docs/development/AutomaticInstrumentation.md#current-limitations)).
Here it'd be even nicer to generate DuckCopy structs instead of
interfaces for the stuff we cannot constraint, but it looked
complicated.
- there is one kind of template type that was not handled, I don't know
why or how it's different from the other, I just did a fix.

see commit list for full list

## Reason for change

generate usable code

## Implementation details

## Test coverage

no tests AFAIK 😬 I used the tool on Avro, comparing with what I ended up
putting in [my
PR](https://github.com/DataDog/dd-trace-dotnet/pull/6882/files) because
it was an instrumentation full of corner cases.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->