Protect against null ref in W3CTraceContextPropagator.cs (#6991)
## Summary of changes

Adds a null check for `TryGetSingleRare` which will prevent
`NullReferenceException`s from being thrown by
`TryGetSingle`/`TryGetSingleRare`.

## Reason for change

Spotted in Error Tracking.

I'm not fully sure what the real world case is that causes this though
for the header to return an `null` value

## Implementation details

Added a null check in `TryGetSingleRare` - the parent `TryGetSingle`
will fall through on `null` to this function so it handles both.

## Test coverage

Added `null` tests. Again, unsure exactly what the realworld cause of it
so pretty synthetic as it doesn't address why the value is null but will
at least not have an exception.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->