[dd-trace] Improved GAC commands (#6995)
## Summary of changes

This PR improves the support in the GAC command, by allowing us to
install, uninstall and get info for multiple versions of an assembly.

## Reason for change

The current support only checks if an assembly is installed in the GAC
no matter version, causing an issue when a previous version of dd-trace
installed an old version of Datadog.Trace in the GAC meaning that we
bailout and don't install the new version, causing a crash when the
application loads.

<img width="1666" alt="image"
src="https://github.com/user-attachments/assets/f4a8b940-6718-470d-bc8c-0a6fc84bb838"
/>

## Implementation details

Now we include more unmanaged Fusion apis to list the version of the
assemblies installed in the GAC.

## Test coverage

Added a new `GacWithMultipleVersions` test.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->