Fix exit code from `GacUninstallCommand` when the assembly is not in the GAC (#7003)
## Summary of changes

Fixes the return code for the `GacUninstallCommand` when the assembly is
not in the GAC

## Reason for change

If the assembly isn't in the GAC, we shouldn't return error when
uninstalling it, because the desired state is correct. This was changed
in #6995 but is an unintended behaviour change

## Implementation details

- Change the exit code `1` -> `0`
- Write a success message instead of a warning
- Fix typo 

## Test coverage

🙈 

## Other details

Currently some of the other "failure" messages are _effectively_ the
same thing, so it might make change to change those too. However, their
behaviour wasn't changed in #6995 so is non-critical