[Tracer] Fix resource-based sampling for `ASP.NET` spans in IIS (#6936)
## Summary of changes

Set's a preliminary `Span.ResourceName` for `ASP.NET` spans in
`OnBeginRequest` as well as in `OnEndRequest` to ensure that
resource-based sampling will work.

## Reason for change

Within `TracingHttpModule` we do not set a `ResourceName` for the span
in `OnBeginRequest`, this is to try and ensure that the resource names
from the parent `ASP.NET` span and it's child span will be equivalent.
We instead set it in the `OnEndRequest`.

This poses an issue though in IIS as we then (commonly) call `Inject`
which ends up making a sampling decision _without_ a
`Span.ResourceName`. What this ends up causing is that resource-based
sampling decisions / rules will never be applied.
Once a sampling decision has been made it is not changed at a later
point.

The following code path gets hit for `ASP.NET` in IIS which causes the
behavior.
```csharp
// Decorate the incoming HTTP Request with distributed tracing headers
// in case the next processor cannot access the stored Scope
// (e.g. WCF being hosted in IIS)
if (HttpRuntime.UsingIntegratedPipeline)
{
    var injectedContext = new PropagationContext(scope.Span.Context, Baggage.Current);
    tracer.TracerManager.SpanContextPropagator.Inject(injectedContext, requestHeaders.Wrap());
}
```

## Implementation details

I've extracted _a part_ of the logic to build the resource name in the
`OnEndRequest` and am using it in the `OnBeginRequest` phase. This will
allow us to have some semblance of a relatively accurate resource name
at the time that we are making the sampling decision.

The part that is different is that we do not in the `OnBeginRequest`
have access to the child span resource name, which is the path ->

```csharp
if (app.Context.Items[SharedItems.HttpContextPropagatedResourceNameKey] is string resourceName
 && !string.IsNullOrEmpty(resourceName))
{
    currentSpan.ResourceName = resourceName;
}
```

So while I _think_ that the resource name computed now will be
relatively accurate there may be instances where the child HTTP span
from it has a different resource name that overrides it.

## Test coverage

Updated MVC4 ASP net tests to have a resource that should be sampled out
(see the snapshot sampling value) based on its `aspnet` span (not the
`MVC`) span resource name.

## Other details

It may be beneficial to tag the original resource name as it may look a
bit different to help determine what resource was used in the sampling
decision (if any), but we can come back to that later.

<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->