Pre-install the .NET SDK in Windows CI runners (#7001)
## Summary of changes

Pre-install the .NET SDKs/runtimes in the Windows VMs

## Reason for change

Currently we're explicitly installing the .NET SDK/runtime in the
Windows stages. This takes time. Seeing as we pre-build the Windows VMs
anyway, we can just pre-install the .NET SDKs there, and set the
`PATH`/`DOTNET_32` variables in the VMs.

## Implementation details

Updated the VMs to pre-install the .NET SDKs, and remove the tasks that
were installing the SDKs in the pipeline.

Note that we still need to install the SDKs in several places:

- When we use the Azure Devops runners - which currently we use for the
smoke tests.
- The macOS stages, because they're hosted runners

## Test coverage

Ran a full test, and all looks good AFAICT.

## Other details

Updated the documentation to include pre-installing the SDKs