Ensure that there is an ActiveScope for AWS SDK Http integration (#7013)
## Summary of changes

Ensure that `InternalActiveScope` isn't `null` as it apparently can be.

## Reason for change

Error tracking alerted on a test app with a `NullReferenceException` for
this.

## Implementation details

Added a `?`

## Test coverage

😬 

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->