Allow specifying `SampleName` filter and comprehensive version testing (#7015)
## Summary of changes

Fix the `BuildSamples` stage so that it respects the `SampleName` and
`perform_comprehensive_testing` CI variables.

## Reason for change

We use these variables to be able to perform testing against _all_
package versions when we make risky modifications to an integration.
After the refactor to `BuildSamples` this functionality was broken.

## Implementation details

Add the variables back

## Test coverage

I'll do a dedicated run to make sure it's working as expected

## Other details

Required for #6989