Add `net8.0` and `net9.0` to trimming nuspec (#6983)
## Summary of changes

Adds `net8.0` and `net9.0` to trimming `nuspec` file.

## Reason for change

Seems strange to have the others declared and was asked if we support
those later ones which we do.

## Implementation details

## Test coverage

## Other details

I'm actually unsure why we have a `nuspec` for this
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->