Fix `ErrorSkipTelemetry` by changing primary file logger to sub-logger config (#7049)
## Summary of changes

This changes the primary file logging configuration to be a sub logging
configuration.

## Reason for change

We added `ErrorSkipTelemetry` in
https://github.com/DataDog/dd-trace-dotnet/pull/7028, but it doesn't
seem to work.
From what I can tell the `Enrich.With(new
RemovePropertyEnricher(LogEventLevel.Error,
DatadogSerilogLogger.SkipTelemetryProperty))` was added to the main
logging configuration and it would hit first and then remove the skip
telemetry property, so now it is it's own sub-logging configuration to
try to not do that.

## Implementation details

Copied the redacted error logging

## Test coverage

Changed previous test to be called "WithoutFile" and added a new one
with a file. The new one fails without the change indicating that the
property is infact removed, and passes with the change.

~~They still pass so I think that is both good and bad as it means we
aren't testing the interaction between the two logging configurations,
but I don't really know how important this is.~~

## Other details

I don't know if this is the right move
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->