Don't fail the IIS instrumentation installation if IIS is not available (#7071)
## Summary of changes

If a server doesn't have IIS installed (or doesn't have a sufficiently
high version) then don't return an error when trying to enable IIS
instrumentation

## Reason for change

Currently, if customers try to enable the IIS instrumentation on servers
which do not have IIS, then the installation fails. This is problematic
for some customers, as it means they need to tailor their installation
script based on the features enabled on the host. As a result, we now
want to not fail the installation in this scenario.

## Implementation details

Just return success if they don't have IIS available

## Test coverage

TODO: Will add a smoke test for this scenario

## Other details

Supersedes:
- https://github.com/DataDog/dd-trace-dotnet/pull/7069

Stacked on: 
- https://github.com/DataDog/dd-trace-dotnet/pull/7063
- https://github.com/DataDog/dd-trace-dotnet/pull/7068