Fix the system-tests parametric tests (#7076)
## Summary of changes

Fix the fact we're not running the parametric tests against the code in
the branch

## Reason for change

For ~6 months, we've not been running the parametric tests against the
code in the branch, we've been running it against the latest release 🙈

## Implementation details

Download the artifact again in the parametric test stages. It's not
required by the other scenarios, because they use the weblog, which
already bakes it in. The parametric tests are different and don't use
the weblog

## Test coverage

Ran a test
[here](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=179057&view=logs&j=fa40fe21-a6c3-5cd1-322e-e17d6fd01752&t=9efec318-bc59-5db9-db95-e03f15f49cd3),
and confirmed it's using the 3.19.0 version (instead of 3.18.0, the
current release) which is what it was using before