Update list of processes excluded from instrumentation (#6599)
## Summary of changes

Adds additional processes to be excluded from tracing.

## Reason for change

- There are some processes that we trace, but don't want to, so added
those.
- Our way of getting the process name was limited to only 15 characters
and didn't work well so had to change that.

## Implementation details

- Added the three `.exe`s from https://github.com/Azure/WindowsVMAgent
- Tries to read the process name from `/proc/self/exe` as it is more
consistent.

## Test coverage

- Added a test for ignored processes
- Added a test for non-ignored processes - we have some already here,
but this is to ensure that we _do_ instrument the test and it isn't just
a broken setup causing the ignored process to not be traced.

## Other details

Comment said to update SSI `requirements.json`, but I looked at them
(the block list) and wasn't really sure what to do as they looked a bit
too different 👀

I've removed all case insensitivity from this as that is another issue
that we saw where process names were of different casing that gets
around the exclusion list. The reason for this was because the
implementation wasn't good enough for all different character sets.

See the commit 14468b5e5ac4903b36ec27d18077dcfcaae61107 for info
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->