Replace fine-grained token usage with app token (#7091)
## Summary of changes

Replaces our current usage of a fine-grained token, with a token based
on a GitHub App

## Reason for change

With recent security changes, we would have to use fine-grained tokens,
and rotate them every 30 days. By using a [trusted GitHub
app](https://github.com/organizations/DataDog/settings/apps/dd-trace-dotnet-ci-bot)
to generate the tokens, we don't have to. Instead, we store a private
key and app ID, and use _that_ to generate a short-lived token.

## Implementation details

In GitHub it's easy enough, as [there's a GitHub
action](https://github.com/actions/create-github-app-token) that handles
it. In Azure DevOps it's more of a pain. You need to:
- [Generate a
JWT](https://docs.github.com/en/apps/creating-github-apps/authenticating-with-a-github-app/generating-a-json-web-token-jwt-for-a-github-app#about-json-web-tokens-jwts)
using the private key and IDs associated with the app
- Make a request with the JWT to retrieve [an installation access
token](https://docs.github.com/en/apps/creating-github-apps/authenticating-with-a-github-app/authenticating-as-a-github-app-installation#generating-an-installation-access-token)
- Use that installation token as the bearer token for subsequent API
requests

To encapsulate that flow, created an Azure DevOps step:
`steps/create-github-app-token` which does all of the above, and writes
the token as a secret, which can be reused in subsequent steps.

One complexity is that this retrieves private key (which contains line
breaks) from an Azure Variable. tl;dr; I had a bunch of issues with
this, so I stored the file with the line breaks replaced with explicit
`\n` characters, which are replaced in the script to generate the "real"
pem file.

## Test coverage

I did some manual testing of the GihHb actions, and I'm pretty sure
they're ok. Eventually got there with Azure DevOps too so 🤞 We'll need
to keep an eye out though

## Other details

Stored the secrets in 1Password, and will update the Secrets doc in
confluence