Fix bug in `dd-dotnet`'s `ExporterSettings` (#7044)
## Summary of changes

- Refactor `ExporterSettingsTests` so that we test the `dd-dotnet`
implementation too
- Run `ExporterSettingsTests` with `dd-dotnet` tests
- Fix bug in `dd-dotnet`'s `ExporterSettings`

## Reason for change

We had a report that the `dd-dotnet` tool was using the wrong port when
_only_ the `DD_AGENT_HOST` value was provided (and `DD_TRACE_AGENT_PORT`
or similar was _not_ provided). Note that this _only_ affected the
`dd-dotnet check process` commands et al. The tracer itself is
functioning correctly:


![image](https://github.com/user-attachments/assets/0b90b36e-c2ff-4077-a99f-fa6f37338856)

The reason for the discrepancy is that `dd-dotnet` uses various "slimmed
down" objects that are compatible with NativeAOT and various other
aspects. There was a bug in this code which was defaulting the port to
`0`.

To make sure there were no other lingering issues, I refactored the
existing `ExporterSettingsTests` file by splitting it in two (much as we
do for `ExporterSettings`) so that we can run the applicable tests
against _both_ of the `ExporterSettings` tests (the `dd-dotnet` version
doesn't care about metrics etc currently, for example).

## Implementation details

- Move a bunch of methods between `ExporterSettingsTests` and
`ExporterSettingsTests.Shared`
- Add some helper constructors etc to `dd-dotnet`'s slimmed-down
`ExporterSettings`
- Fix bug in `dd-dotnet`'s slimmed-down `ExporterSettings`

## Test coverage

More now

## Other details