[Dynamic Instrumentation] DEBUG-4063 Add WCF support for dynamic span tags (#7097)
## Summary of Changes
This PR adds a way to get the active scope in WCF applications.

## Reason for Change
We have seen that in some cases we cannot get the active scope through
the `ScopeManager`.

## Implementation Details
When an active scope does not exist, and we are in .NET Framework, we
try to get the scope from the current action context.